backend/pyproject.toml
[tool.poetry]
name = "own-prj"
version = "0.1.0"
description = ""
authors = ["Bezizvestnyi <chikovm@bk.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
django-environ = "^0.12.0"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
